src/cleanup-manager/requirements.txt
# =============================================================================
# Cleanup Manager - Python Dependencies
# =============================================================================

# Scheduler (matches the version in CS-GitHubBackup)
APScheduler==4.0.0a6

# Configuration via type-safe ENV parsing
pydantic>=2.12.0
pydantic-settings>=2.12.0

# Console UI
rich>=14.0.0

# JWT signing for GitHub App auth (RS256)
PyJWT[crypto]>=2.10.0
